all ints changed to long

+++ pickANumberExercise/pickANumberExercise/Program.cs
@@ -12,17 +12,17 @@ namespace pickANumberExercise
         {
             Console.WriteLine("Write any number: ");
             string numResponse = Console.ReadLine();
-            int inputNum = Convert.ToInt32(numResponse);
-            int multiply = inputNum * 50;
+            long inputNum = Convert.ToInt64(numResponse);
+            long multiply = inputNum * 50;
             Console.WriteLine(inputNum + " multiplied by 50 is " + multiply);
-            int add = inputNum + 25;
+            long add = inputNum + 25;
             Console.WriteLine(inputNum + " added to 25 is " + add);
             double divide = inputNum / 12.5;
             Console.WriteLine(inputNum + "divided by 12.5 is " + divide);
             bool compare = inputNum > 50;
             Console.WriteLine(inputNum + " is greater than 50?");
             Console.WriteLine(compare);
-            int remainder = inputNum % 7;
+            long remainder = inputNum % 7;
             Console.WriteLine(inputNum + " divided by 7 has a remainder of " + remainder);
             Console.Read();
         }
